+++ .idea/gradle.xml
@@ -10,7 +10,7 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/SpeechDrawable" />
+            <option value="$PROJECT_DIR$/LibrarySpeech" />
             <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
