+++ LibrarySpeech/src/main/AndroidManifest.xml
