+++ LibrarySpeech/src/main/java/com/newolf/libraryspeech/SpeechDrawable.java
