+++ app/src/androidTest/java/com/newolf/speechdrawabledemo/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.newolf.speechdrawable
+package com.newolf.speechdrawabledemo
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
