+++ app/src/test/java/com/newolf/speechdrawabledemo/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.newolf.speechdrawable
+package com.newolf.speechdrawabledemo
 
 import org.junit.Test
