+++ settings.gradle
@@ -1,3 +1,3 @@
-include ':SpeechDrawable'
+include ':LibrarySpeech'
 include ':app'
 rootProject.name = "SpeechDrawable"
